+++ packages/nest/src/open-feature.module.ts
@@ -7,11 +7,21 @@ import {
   Provider as NestProvider,
   ExecutionContext,
 } from '@nestjs/common';
-import { Client, OpenFeature, Provider, EvaluationContext } from '@openfeature/server-sdk';
+import {
+  Client,
+  Hook,
+  OpenFeature,
+  Provider,
+  EvaluationContext,
+  ServerProviderEvents,
+  EventHandler,
+  Logger,
+} from '@openfeature/server-sdk';
 import { ContextFactory, ContextFactoryToken } from './context-factory';
 import { APP_INTERCEPTOR } from '@nestjs/core';
 import { AsyncLocalStorageTransactionContext } from './evaluation-context-propagator';
 import { EvaluationContextInterceptor } from './evaluation-context-interceptor';
+import { ShutdownService } from './shutdown.service';
 
 /**
  * OpenFeatureModule is a NestJS wrapper for OpenFeature Server-SDK.
@@ -20,7 +30,18 @@ import { EvaluationContextInterceptor } from './evaluation-context-interceptor';
 export class OpenFeatureModule {
   static forRoot({ useGlobalInterceptor = true, ...options }: OpenFeatureModuleOptions): DynamicModule {
     OpenFeature.setTransactionContextPropagator(new AsyncLocalStorageTransactionContext());
-    const providers: NestProvider[] = [];
+
+    if (options.logger) {
+      OpenFeature.setLogger(options.logger);
+    }
+
+    if (options.hooks) {
+      OpenFeature.addHooks(...options.hooks);
+    }
+
+    options.handlers?.forEach(([event, handler]) => {
+      OpenFeature.addHandler(event, handler);
+    });
 
     const clientValueProviders: NestFactoryProvider<Client>[] = [
       {
@@ -43,27 +64,27 @@ export class OpenFeatureModule {
       });
     }
 
-    providers.push(...clientValueProviders);
+    const nestProviders: NestProvider[] = [ShutdownService];
+    nestProviders.push(...clientValueProviders);
 
     const contextFactoryProvider: ValueProvider = {
       provide: ContextFactoryToken,
       useValue: options?.contextFactory,
     };
-
-    providers.push(contextFactoryProvider);
+    nestProviders.push(contextFactoryProvider);
 
     if (useGlobalInterceptor) {
       const interceptorProvider: ClassProvider = {
         provide: APP_INTERCEPTOR,
         useClass: EvaluationContextInterceptor,
       };
-      providers.push(interceptorProvider);
+      nestProviders.push(interceptorProvider);
     }
 
     return {
       global: true,
       module: OpenFeatureModule,
-      providers,
+      providers: nestProviders,
       exports: [...clientValueProviders, ContextFactoryToken],
     };
   }
@@ -85,6 +106,26 @@ export interface OpenFeatureModuleOptions {
   providers?: {
     [providerName: string]: Provider;
   };
+  /**
+   * Global {@link Logger} for OpenFeature.
+   * @see {@link OpenFeature#setLogger}
+   */
+  logger?: Logger;
+  /**
+   * Global {@link EvaluationContext} for OpenFeature.
+   * @see {@link OpenFeature#setContext}
+   */
+  context?: EvaluationContext;
+  /**
+   * Global {@link Hook Hooks} for OpenFeature.
+   * @see {@link OpenFeature#addHooks}
+   */
+  hooks?: Hook[];
+  /**
+   * Global {@link EventHandler EventHandlers} for OpenFeature.
+   * @see {@link OpenFeature#addHandler}
+   */
+  handlers?: [ServerProviderEvents, EventHandler][];
   /**
    * The {@link ContextFactory} for creating an {@link EvaluationContext} from Nest {@link ExecutionContext} information.
    * This could be header values of a request or something similar.