+++ packages/react/CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## [0.1.0-experimental](https://github.com/open-feature/js-sdk/compare/react-sdk-v0.0.6-experimental...react-sdk-v0.1.0-experimental) (2024-01-18)
+
+
+### ⚠ BREAKING CHANGES
+
+* remove generic hook, add specific type hooks ([#766](https://github.com/open-feature/js-sdk/issues/766))
+
+### ✨ New Features
+
+* remove generic hook, add specific type hooks ([#766](https://github.com/open-feature/js-sdk/issues/766)) ([d1d02fa](https://github.com/open-feature/js-sdk/commit/d1d02fa59de5b5b1b8866c0b5d3de1a5bc0c5a04))
+
+
+### 🧹 Chore
+
+* fix react-sdk REAMDE example, add missing `EvaluationContext` ([#762](https://github.com/open-feature/js-sdk/issues/762)) ([1e13333](https://github.com/open-feature/js-sdk/commit/1e1333381909b790d0c4fc7590613b2ae6f1aa2e))
+
+## [0.0.6-experimental](https://github.com/open-feature/js-sdk/compare/react-sdk-v0.0.5-experimental...react-sdk-v0.0.6-experimental) (2024-01-11)
+
+
+### ✨ New Features
+
+* suspense support, client scoping, context-sensitive re-rendering ([#759](https://github.com/open-feature/js-sdk/issues/759)) ([8f01ead](https://github.com/open-feature/js-sdk/commit/8f01ead29104d122b8126d2fee97c98556091344))
+
+
+### 🧹 Chore
+
+* fix React SDK README.md to use the resolved value instead of the resolution details ([#691](https://github.com/open-feature/js-sdk/issues/691)) ([2d1b8eb](https://github.com/open-feature/js-sdk/commit/2d1b8ebfb187db92db02ea36fa6d6ca291591b18))
+* react-sdk | downgrading react peer dependency | react 18.0.0 -&gt; 16.8.0 ([#742](https://github.com/open-feature/js-sdk/issues/742)) ([2c864e4](https://github.com/open-feature/js-sdk/commit/2c864e46ccecd6d8825738f30a2d098dc66e26cf))
+
 ## [0.0.5-experimental](https://github.com/open-feature/js-sdk/compare/react-sdk-v0.0.4-experimental...react-sdk-v0.0.5-experimental) (2023-11-27)